LPS-79385 announcements-uad - renames *Anonymizer classes and updates references
