+++ modules/apps/collaboration/announcements/announcements-uad/src/main/java/com/liferay/announcements/uad/anonymizer/AnnouncementsEntryUADAnonymizer.java
@@ -14,16 +14,17 @@
 
 package com.liferay.announcements.uad.anonymizer;
 
-import com.liferay.announcements.kernel.model.AnnouncementsFlag;
-import com.liferay.announcements.kernel.service.AnnouncementsFlagLocalService;
+import com.liferay.announcements.kernel.model.AnnouncementsEntry;
+import com.liferay.announcements.kernel.service.AnnouncementsEntryLocalService;
 import com.liferay.announcements.uad.constants.AnnouncementsUADConstants;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.user.associated.data.anonymizer.DynamicQueryUADEntityAnonymizer;
-import com.liferay.user.associated.data.anonymizer.UADEntityAnonymizer;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.user.associated.data.anonymizer.DynamicQueryUADAnonymizer;
+import com.liferay.user.associated.data.anonymizer.UADAnonymizer;
 import com.liferay.user.associated.data.util.UADAnonymizerHelper;
 
-import java.util.Collections;
+import java.util.Arrays;
 import java.util.List;
 
 import org.osgi.service.component.annotations.Component;
@@ -34,44 +35,50 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(
 	immediate = true,
-	property = "model.class.name=" + AnnouncementsUADConstants.CLASS_NAME_ANNOUNCEMENTS_FLAG,
-	service = UADEntityAnonymizer.class
+	property = "model.class.name=" + AnnouncementsUADConstants.CLASS_NAME_ANNOUNCEMENTS_ENTRY,
+	service = UADAnonymizer.class
 )
-public class AnnouncementsFlagUADEntityAnonymizer
-	extends DynamicQueryUADEntityAnonymizer<AnnouncementsFlag> {
+public class AnnouncementsEntryUADAnonymizer
+	extends DynamicQueryUADAnonymizer<AnnouncementsEntry> {
 
 	@Override
-	public void autoAnonymize(AnnouncementsFlag announcementsFlag, long userId)
+	public void autoAnonymize(
+			AnnouncementsEntry announcementsEntry, long userId)
 		throws PortalException {
 
-		announcementsFlag.setUserId(_uadAnonymizerHelper.getAnonymousUserId());
+		User anonymousUser = _uadAnonymizerHelper.getAnonymousUser();
 
-		_announcementsFlagLocalService.updateAnnouncementsFlag(
-			announcementsFlag);
+		announcementsEntry.setUserId(anonymousUser.getUserId());
+		announcementsEntry.setUserName(anonymousUser.getFullName());
+
+		_announcementsEntryLocalService.updateAnnouncementsEntry(
+			announcementsEntry);
 	}
 
 	@Override
-	public void delete(AnnouncementsFlag announcementsFlag) {
-		_announcementsFlagLocalService.deleteFlag(announcementsFlag);
+	public void delete(AnnouncementsEntry announcementsEntry) {
+		_announcementsEntryLocalService.deleteAnnouncementsEntry(
+			announcementsEntry);
 	}
 
 	@Override
 	public List<String> getNonanonymizableFieldNames() {
-		return Collections.emptyList();
+		return Arrays.asList("content", "title");
 	}
 
 	@Override
 	protected ActionableDynamicQuery doGetActionableDynamicQuery() {
-		return _announcementsFlagLocalService.getActionableDynamicQuery();
+		return _announcementsEntryLocalService.getActionableDynamicQuery();
 	}
 
 	@Override
 	protected String[] doGetUserIdFieldNames() {
-		return AnnouncementsUADConstants.USER_ID_FIELD_NAMES_ANNOUNCEMENTS_FLAG;
+		return AnnouncementsUADConstants.
+			USER_ID_FIELD_NAMES_ANNOUNCEMENTS_ENTRY;
 	}
 
 	@Reference
-	private AnnouncementsFlagLocalService _announcementsFlagLocalService;
+	private AnnouncementsEntryLocalService _announcementsEntryLocalService;
 
 	@Reference
 	private UADAnonymizerHelper _uadAnonymizerHelper;
